pyproject.toml
# line lengths:
#   - Excessive line lengths aren't good. But rather then automatically
#     breaking them and turning into vertical blocks we want to see the
#     nastiness sticking out prominently.
#
# black:
#   - If the question is "Black?" the answer is "No!"
#   - We use autopep8 to confirm compliance with PEP8, going beyond that
#     is not necessary. If you want to apply black, then fine, go ahead.
#     But we're not going to enforce or encourage it.

# https://github.com/PyCQA/isort
# https://pycqa.github.io/isort/
[tool.isort]
# use the black profile but override "multi_line_output"
profile           = "black"
line_length       = 250
multi_line_output = 7
src_paths         = ["src", "tests"]

# https://pypi.org/project/autopep8/
[tool.autopep8]
max_line_length = 250
ignore          = ["E221", "E241"]
in-place = true

# https://python-poetry.org/docs/basic-usage/
[tool.poetry]
name = "feeph-i2c-emc2101"
version = "0.0.0"
description = "library for EMC2101 fan control chip"
authors = [
    "Feeph Aifeimei <55798703+feeph@users.noreply.github.com>",
]
keywords = [
    "emc2101",
    "i2c",
    "smbus",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: System :: Hardware :: Inter-Integrated Circuit",
    "Typing :: Typed",
]
readme = "README.md"
license = "GPL-3.0-or-later"
#homepage =
repository = "https://github.com/feeph/libi2c-emc2101-python/"
#documentation =
packages = [
    {include = "i2c", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.11.0"
adafruit-board-toolkit = "^1.1.0"
gpiod = "^2.1.0"
adafruit-circuitpython-busdevice = "^5.2.0"
pyyaml = "^6.0.0"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/feeph/libi2c-emc2101-python/issues"

[build-system]
requires = [
    "poetry-core"
]
build-backend = "poetry.core.masonry.api"
